Move session start time badge to its own component
